chat_engine/requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
dj-rest-auth==6.0.0
Django==5.1.1
django-allauth==65.0.2
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jwt==1.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.2.1
python-decouple==3.8
requests==2.32.3
service-identity==24.1.0
sqlparse==0.5.1
tomli==2.0.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.3
zope.interface==7.0.3
